add generic deflect screen and remove Emergency and Glass damage
